Turtle pairing frontend and widen IDs

+++ TurtlePublic.Sql/Tables/Event.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Event] (
-    [Id]        INT          IDENTITY (1, 1) NOT NULL,
+    [Id]        BIGINT          IDENTITY (1, 1) NOT NULL,
     [Type]      VARCHAR (50) NOT NULL,
     [Timestamp] DATETIME     CONSTRAINT [DF_Event_Timestamp] DEFAULT (getdate()) NOT NULL,
     [CCType]    VARCHAR (50) NOT NULL,
